file updated

+++ backend/controllers/cartController.js
@@ -5,7 +5,7 @@ const { CartModel } = require("../models/cartModel");
 const addNewItemToCart = async (req, res) => {
   const data = req.body;
   const userId = req.body.userId;
-  const productId = req.body.productId;
+  const productId = req.body.product;
   try {
     const isProductPresentInCart = await CartModel.findOne({
       product: productId,
@@ -13,7 +13,10 @@ const addNewItemToCart = async (req, res) => {
     });
     if (isProductPresentInCart)
       return res.status(409).json({ message: "Already in cart!" });
-    const newCartItem = CartModel({ ...data });
+    const newCartItem = CartModel({
+      user: userId,
+      product: productId,
+    });
     await newCartItem.save();
     return res.status(200).json({ message: "Added to cart" });
   } catch (err) {
@@ -28,6 +31,12 @@ const deleteCartItem = async (req, res) => {
   const userId = req.body.userId;
   const cartId = req.params.cartId;
   try {
+    const isItemPresent =
+      (await CartModel.findOne({ user: userId, _id: cartId })) || null;
+    if (!isItemPresent)
+      return res
+        .status(404)
+        .json({ message: "You are not allow to delete this item." });
     const deletedItem = await CartModel.deleteOne({
       user: userId,
       _id: cartId,
@@ -46,6 +55,12 @@ const updateQtyOfCartItem = async (req, res) => {
   const userId = req.body.userId;
   const cartId = req.params.cartId;
   try {
+    const isItemPresent =
+      (await CartModel.findOne({ user: userId, _id: cartId })) || null;
+    if (!isItemPresent)
+      return res
+        .status(404)
+        .json({ message: "You are not allow to update this item." });
     const updateCart = await CartModel.findOneAndUpdate(
       { user: userId, _id: cartId },
       {
@@ -68,8 +83,11 @@ const updateQtyOfCartItem = async (req, res) => {
 // get all cart item of a particular user.
 const getCartItems = async (req, res) => {
   const userId = req.body.userId;
+  console.log("get cart Called: ", userId);
   try {
-    const cartItems = await CartModel.find({ user: userId });
+    const cartItems = await CartModel.find({ user: userId })
+      .populate("product")
+      .populate("user", "email name role");
     res.status(200).json({ message: "you cart items", cart: cartItems });
   } catch (err) {}
 };
